nix-setup: handle nix upgrades
Since with this method of installation, nix cannot upgrade itself,
the setup service should check whether the current version is lower
than the version of the rpm package.
If that is the case, we want to set the version from the archive as the
active version, but also keep the rest of the nix store and profiles
intact. If I understand `nix-store --dump-db/--load-db` correctly, this
little switching out of the db files to import from a temporarily active
db that knows the new nix version is necessary.

Important to rerun the SELinux stuff after performing this type of update
otherwise the daemon systemd unit will fail.